+++ src/aws_control_tower_manifest_builder/aws_control_tower_manifest_builder.py
@@ -23,13 +23,16 @@ def main(args: None):
 
     if args:
         output_manifest_file = os.path.join(args.output, "manifest.yaml")
-        default_region = args.default_region
 
     final_report = PrettyTable(["Type", "Successes", "Failures", "Totals"])
 
     log.info("Processing CF Templates from %s", args.input_cf)
     cf_resources, cf_successes, cf_failures = loop_through_files(
-        args.input_cf, manifest_input.CfTemplate, default_region
+        args.input_cf,
+        manifest_input.CfTemplate,
+        args.default_region,
+        args.metadata_name,
+        args.enforce_account_number_only,
     )
     final_report.add_row(
         [
@@ -42,7 +45,11 @@ def main(args: None):
 
     log.info("Processing SCPs from %s", args.input_scp)
     scp_resources, scp_successes, scp_failures = loop_through_files(
-        args.input_scp, manifest_input.Scp, default_region
+        args.input_scp,
+        manifest_input.Scp,
+        args.default_region,
+        args.metadata_name,
+        args.enforce_account_number_only,
     )
     final_report.add_row(
         [
@@ -63,7 +70,7 @@ def main(args: None):
     )
 
     data = {"resources": cf_resources + scp_resources}
-    data["region"] = default_region
+    data["region"] = args.default_region
     data["schema_version"] = args.schema_version
 
     loaded_environment = Environment(loader=FileSystemLoader(JINJA_MANIFEST_PATH))
@@ -85,7 +92,11 @@ def main(args: None):
 
 
 def loop_through_files(
-    path: str, manifest_type: manifest_input, default_region: str
+    path: str,
+    manifest_type: manifest_input,
+    default_region: str,
+    metadata_name: str,
+    enforce_account_number_only: bool,
 ) -> list:
     """
     Loop through path and create list of either CfTemplate resources or SCPs
@@ -93,6 +104,7 @@ def loop_through_files(
     Parameters:
     path(str): location of CFN or SCP
     manifest_type(manifest_input): which current manifest_type [SCP or CFN] thats beeing looked for.
+    metadata_name(str): name in of the metadata inside the yaml file
 
     Return:
     [resources(list), successes(int), failures(int)]
@@ -103,7 +115,9 @@ def loop_through_files(
     for filename in sorted(os.listdir(path)):
         path_to_file = os.path.join(path, filename)
         if os.path.isfile(path_to_file) and ".yaml" in filename:
-            new_input = manifest_type(path_to_file, default_region)
+            new_input = manifest_type(
+                path_to_file, default_region, metadata_name, enforce_account_number_only
+            )
             if new_input.metadata_dict and not new_input.error:
                 log.info("Processed .. %s", path_to_file)
                 resources.append(new_input.metadata_dict)